Refactored crown components to better encapsulate methods related to buttons.
