+++ src/components/deleteButton.vue
@@ -1,8 +1,11 @@
 <template>
   <crown-button
+    :title="$t('Delete')"
+    role="menuitem"
     id="delete-button"
     aria-label="Delete this node"
-    v-on="$listeners"
+    @click="isPoolLane ? removePoolLaneShape() : removeShape()"
+    v-b-tooltip.hover.viewport.d50
   >
     <img
       :src="trashIcon"
@@ -21,10 +24,50 @@ import CrownButton from '@/components/crownButton';
 
 export default {
   components: { CrownButton },
+  props: { graph: Object, shape: Object, node: Object },
   data() {
     return {
       trashIcon,
     };
+  },
+  computed: {
+    isPoolLane() {
+      return this.node.type === 'processmaker-modeler-lane';
+    },
+  },
+  methods: {
+    removeShape() {
+      this.graph.getConnectedLinks(this.shape).forEach(shape => this.$emit('remove-node', shape.component.node));
+      this.shape.getEmbeddedCells({ deep: true }).forEach(cell => {
+        if (cell.component) {
+          this.graph.getConnectedLinks(cell).forEach(shape => this.$emit('remove-node', shape.component.node));
+          this.shape.unembed(cell);
+          this.$emit('remove-node', cell.component.node);
+        }
+      });
+
+      this.$emit('remove-node', this.node);
+    },
+    removePoolLaneShape() {
+      this.$emit('remove-node', this.node);
+
+      const poolComponent = this.node.pool.component;
+      const sortedLanes = poolComponent.sortedLanes();
+
+      if (sortedLanes.length === 2) {
+        /* Do not allow pool with only one lane;
+         * if removing 2nd last lane, remove the other lane as well */
+        this.$emit('remove-node', sortedLanes.filter(lane => lane !== this.shape)[0].component.node);
+        return;
+      }
+
+      if (this.shape === sortedLanes[sortedLanes.length - 1]) {
+        poolComponent.fillLanes(this.shape, 'top-right', true);
+        return;
+      }
+
+      poolComponent.fillLanes(this.shape, 'bottom-right', true);
+    },
   },
 };
 </script>
