+++ src/components/messageFlowButton.vue
@@ -1,9 +1,13 @@
 <template>
   <crown-button
+    v-if="validMessageFlowSources.includes(node.type)"
+    v-b-tooltip.hover.viewport.d50
+    :title="$t('Message Flow')"
     id="message-flow-button"
     aria-label="Create a message flow"
     :src="messageFlow"
-    v-on="$listeners"
+    role="menuitem"
+    @click="addMessageFlow"
   />
 </template>
 <script>
@@ -12,10 +16,35 @@ import CrownButton from '@/components/crownButton';
 
 export default {
   components: { CrownButton },
+  props: ['node', 'moddle', 'shape'],
   data() {
     return {
       messageFlow,
+      validMessageFlowSources: [
+        'processmaker-modeler-start-event',
+        'processmaker-modeler-end-event',
+        'processmaker-modeler-task',
+        'processmaker-modeler-pool',
+        'processmaker-modeler-intermediate-message-catch-event',
+      ],
     };
+  },
+  methods: {
+    addMessageFlow(cellView, evt, x, y) {
+      this.$emit('show-crown', false);
+
+      const messageFlowDefinition = this.moddle.create('bpmn:MessageFlow', {
+        name: '',
+        sourceRef: this.shape.component.node.definition,
+        targetRef: { x, y },
+      });
+
+      this.$emit('add-node', {
+        type: 'processmaker-modeler-message-flow',
+        definition: messageFlowDefinition,
+        diagram: this.moddle.create('bpmndi:BPMNEdge'),
+      });
+    },
   },
 };
 </script>
