--- src/components/poolLaneCrownConfig.vue
@@ -1,30 +1,0 @@
-<script>
-import crownConfig from '@/components/crownConfig';
-
-export default {
-  extends: crownConfig,
-  methods: {
-    removeShape() {
-      this.$emit('remove-node', this.node);
-
-      const poolComponent = this.node.pool.component;
-      const sortedLanes = poolComponent.sortedLanes();
-
-      if (sortedLanes.length === 2) {
-        /* Do not allow pool with only one lane;
-         * if removing 2nd last lane, remove the other lane as well */
-        this.$emit('remove-node', sortedLanes.filter(lane => lane !== this.shape)[0].component.node);
-        return;
-      }
-
-      if (this.shape === sortedLanes[sortedLanes.length - 1]) {
-        poolComponent.fillLanes(this.shape, 'top-right', true);
-        return;
-      }
-
-      poolComponent.fillLanes(this.shape, 'bottom-right', true);
-    },
-  },
-};
-
-</script>
